+++ HowManyCities/Models/Ratio.swift
@@ -1,0 +1,13 @@
+//
+//  Ratio.swift
+//  HowManyCities
+//
+//  Created by Geoffrey Liu on 5/9/22.
+//
+
+import Foundation
+
+struct Ratio: Hashable {
+  let numerator: Int
+  let denominator: Int
+}
