+++ HowManyCities/Stats/Subviews/Item Renderers/CityPopulationRenderer.swift
@@ -8,14 +8,6 @@
 import UIKit
 
 final class CityPopulationRenderer: ItemRenderer {
-  func render(_ item: City) -> UIView? {
-    let label = UILabel().autolayoutEnabled
-    label.numberOfLines = 2
-    label.attributedText = string(item)
-    
-    return label
-  }
-  
   func string(_ item: City) -> NSAttributedString {
     let mas: NSMutableAttributedString
     if let countryFlag = item.countryFlag {
@@ -36,33 +28,3 @@ final class CityPopulationRenderer: ItemRenderer {
 }
 
 
-final class MultiCityPopulationRenderer: ItemRenderer {
-  typealias Item = [City]
-  
-  func render(_ item: [City]) -> UIView? {
-    nil
-  }
-  
-  func string(_ item: [City]) -> NSAttributedString {
-    let population = item.first?.population ?? 0
-    
-    let mas: NSMutableAttributedString
-    let firstFewCitiesNames = item.prefix(3).map(by: \.nameWithStateAbbr).joined(separator: "; ")
-    let flagString: String = item.flags.isEmpty ? "" : "\(item.flags) "
-    mas = .init(string: "\(flagString)\(firstFewCitiesNames)")
-    if item.count > 3 {
-      let numCitiesRemaining = item.count - 3
-      // TODO: Proper pluralization
-      let remainingString = numCitiesRemaining > 1 ? "…and \(numCitiesRemaining.commaSeparated) others" : "…and \(numCitiesRemaining) other"
-      mas.append(.init(string: "\n\t\(remainingString)  ", attributes: [.font: UIFont.italicSystemFont(ofSize: UIFont.labelFontSize),
-                                                                      .foregroundColor: UIColor.label.withAlphaComponent(0.8)]))
-    } else {
-      mas.append(.init(string: "  "))
-    }
-    
-    mas.append(.init(string: population.abbreviated, attributes: [.font: UIFont.systemFont(ofSize: UIFont.smallSystemFontSize),
-                                                                       .foregroundColor: UIColor.systemGray]))
-    
-    return .init(attributedString: mas)
-  }
-}
